CB-70 Book a cleaning (+ minor bugs)
Added a service-method and the controller-method for requesting new cleaning jobs. Added the necessary DTOs, custom exceptions and custom query-methods to do the job as the user story is described.

Had to change the CustomerEntity to correct a bug in where the phone number wasn't persisted as expected. It works best as a String rather than a Integer.

Added some initData for easier testing.