CB-70 Book a cleaning (feature)
Added the feature of sending an email when a customer requests a new job. The email is sent via an Gmail account set up with SMTP.

The application.yml-file is updated to include the settings needed for connecting to the Gmail-account. The credentials needs to be provided via env.properties for the service to work.

pom.xml-file is updated with the necessary dependencies for sending emails.