CB-81 Assign cleaner (bug)
Updated the validation-method to include a check for the list of employee ids. Previously the method only checked each id in the list for a valid input value and id - but didn't check that the actual list was provided with the request - which caused the server to incorrectly respond with a 500-status code if the list were missing instead of a proper 404 with a message.