CREATED: updateEmployee/getAllCleaners in AdminController
Created a new end-point for fetching all cleaners and an end-point for handling updates of employees in the AdminController.

The update-method could be re-used for updating both admins and cleaners. It makes sure the request is sent by an admin and checks for new values and then update the employee with the new values. It doens't matter at this stage if the employee is a cleaner or an admin.