REFACTOR: Update-password method
Updated the method to allow the server to check if the old password is correct before proceeding with the change. This is achieved by extracting the username from the JWT and calling the login-method, and then calling the API-call for updating the password if the login-call is successful. If the password is incorrect the method will return a 403-FORBIDDEN.

Changed the return code to 204 if successful.

BUGFIX: Added the client_cleaner role as well to the end-point for updating employees since the cleaner was forbidden from updating his or her data.