REFACTOR: Cancel job request as customer
Rewrote the method to extract the customerId from the header instead of requiring the client to send this in the body as well.

Deleted code that no longer is needed due to the role-based authentication with the JWT generated from Keycloak. Only a user with the role of customer can access this end-point.