REFACTOR: MailSenderService to allow for HTML-formatting (WIP)
Added the MimeMessage class and the MimeMessageHelper to allow for formatting the email with HTML. Right now it just send a generic message for each method, and we have to adapt the methods to request an HTML-snippet from the client before sending an email.

The logic works and an email is sent with HTML-formatting.

Updated the SecurityFilterChainConfig to allow for POST-request to the contact-us form end-point as well.